requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.18.0
pillow==10.1.0
numpy>=1.26.0
requests==2.31.0
opencv-python==4.9.0.80
